Update patch set 1
Patch Set 1:

Thanks for the fix. I believe this bug should already be fixed by ag/1740147, internal bug b/32156769 (which is in N-MR2) which prevents the media buffer group from growing beyond the growth limit.

I think this change is interesting, which is not to put the transfer buffer on the media buffer group at all and let mIndexCache hold a ref to the IMemory.

Patch-set: 1
Reviewer: Gerrit User 1040213 <1040213@85c56323-6fa9-3386-8a01-6480fb634889>