Avoid OMXNodeInstance be used after freeNode
If OMXNodeInstance::freeNode was called before OMX::CallbackDispatcher
was destructed, OMXNodeInstance will delete itself in the function of
freeNode(). At the same time, there may be unhandled message in CallbackDispatcher's
mQueue.  CallbackDispatcher::dispatch may be called before destructor, which
will call OMXNodeInstance::handleMessage().  In this case, OMXNodeInstance
has been deleted, when try to getGraphicBufferSource, the value of
mGraphicBufferSourceLock is uncertainty, which may result in deadlock.

Test: Repeatedly playback videos with MediaPlayer may reproduct this issue.
Change-Id: Icd06bce98b1f19bc2efb827360f52db5f0b1b4e2
Signed-off-by: zengjing <zengjing@xiaomi.com>