libmedia: Improve charset detection.
Tags are first put together before being detected. When different tags
are encoded in different charset, the result could be very bad. So when
the best result is not good enough, we should not accept it. We give it
another chance by detecting tags separately.

For the test samples we collected, 6 songs out of 13 are now displayed
correctly.

Test: transfer mp3 to devices and see if the tags' charsets are detected
correctly.
Signed-off-by: caozhiyuan <cao.zhiyuan@zte.com.cn>

Change-Id: I90d11a612b9f9e8896696df85635c4a46d067f09