M3UParser should return err when missing media URIs
When parsing an malformed m3u8 file, where media URI was missed, the
mItems.size() was 0. In this case, if mFirstSeqNumber was not set,
the sum of mFirstSeqNumber and mItems.size() will be 0. When executing
the following code:
    mLastSeqNumber = mFirstSeqNumber + mItems.size() - 1;
The reslut turn unsigned after the addition of mFirstSeqNumber and
mItems.size(), in this case, overflow happened by minus 1. The mediaserver
will abort. (This issue only happen in Android N, for unsigned-integer-overflow
and signed-integer-overflow flag was set when compile in Android N).

issues:38185944

Test: Found an malformed m3u8 steam without media URI. Playback it with MediaPlayer.
Change-Id: I7490ab8d3b3528921a1b9939c45257edbba737e4