Allow edit list atom to appear earlier in the file
While processing the 'elst' atom, MPEG4Extractor was expecting duration
and sample rate to already be known. This implies an ordering of atoms
that is not required by the spec, and indeed some muxers may emit the
'elst' atom before the atoms that define sample rate and duration.
This change caches the values from the 'elst' atom until a track has
been completely processed, allowing the 'elst' atom to appear first.

Bug: 63390273
Test: decoded test files, ran CTS
Change-Id: I30abca70ead7990eb8a7d19a2e4b31212b9bc311