Update patch set 1
Patch Set 1:

The problem seems to be the use of binutils nm on LLVM IR (as opposed to llvm-nm, but llvm-nm breaks on gcc LTO bytecode, so that may not be a 100% solution either...)

In OpenMandriva (where we have even more of a mixed compiler setup than in AOSP), we solved this problem by replacing nm with a wrapper script:

https://github.com/OpenMandrivaAssociation/binutils/blob/master/nm

Essentially this looks at the files and invokes binutils-nm or llvm-nm depending on what the input file is. Maybe this or something similar needs to go into AOSP nm.

Patch-set: 1
Reviewer: Gerrit User 1010121 <1010121@85c56323-6fa9-3386-8a01-6480fb634889>