Update patch set 2
Patch Set 2:

I agree that the expected behavior of a copy in Windows is to copy recursively. However, the MTP specification doesn't state that the copyObject operation should copy the children of the copied object as well. We think it is a bug in Windows, and Windows should be sending additional copyObject requests for the children when it is trying to copy a folder object.

Patch-set: 2
Reviewer: Gerrit User 1113594 <1113594@85c56323-6fa9-3386-8a01-6480fb634889>