Fix missing path handling after the deletion
During the deletion of a folder, after unlinking a file or removing a
sub-folder, the deleted file/sub-folder name should be cut off from
the path string. Otherwise, the next file/sub-folder name will be
directly appended to the path string with current file/sub-folder name.
Subsequent deletion will fail because of the wrong path.

Test: Delete a folder containing multiple sub-folders and files by
MTP interface, and check the status of the deleted file/sub-folders
with adb shell

Change-Id: I1768a3ab6cb77ede963a9c0fab2df82736eaf8b7
Signed-off-by: kyle_tso <kyle_tso@htc.com>