Update patch set 1
Patch Set 1:

> This seems a very specific fix for just the ringtone case.
 > Are you sure you don't want to mute everything when playing a
 > shutter sound when the sound is enforced by regulation?

I'm intentionally fixing only for sonification case since we got complaints from some operator only for this. Actual use case is receiving incoming call during video recording when headset is connected. In this case, video recording stopping sound and ringtone are mixed and heard from speaker so the sound is not clear. Other stream type will not be heard from speaker in this case so it's not a big problem.

Patch-set: 1
Reviewer: Gerrit User 1243029 <1243029@85c56323-6fa9-3386-8a01-6480fb634889>