VT: StagefrightRecorder: Added a function for TMMBR
EncodingBitrate was not updated even IMSService
called a setEncodingBitrate() while VideoRecorder is in running state.

Because encoding bitrate can't be updated dynamically in running state.
This patch makes that StagefrightRecorder can set a new encoding bitrate
to satisfy TMMBR requirement as per RFC 5104.

Regular I frames overloads on the network so we should consider about it.
Discounted encoding bitrate will be margins for the overloads.

Change-Id: Ied820dec5bda36bfe590cf46ec981e0b5fc4c4ee
Signed-off-by: Byeongjo Park <bjo.park@samsung.com>
Signed-off-by: Kim Sungyeon <sy85.kim@samsung.com>