Update patch set 1
Patch Set 1:

Hi Eric, Andy
Thank you very much for your review.
One of our effect has a filter to process the data in the frequency-domain so that the effect has an internal buffer.
When switching to next/previous track, we want to flush the internal buffer otherwise the old data in the internal buffer is used unintentionally then it causes the noise.
As Eric pointed out, current implementation is also flushing when transitioning from one song to the next so let me reconsider the proper solution.

Patch-set: 1
Reviewer: Gerrit User 1243029 <1243029@85c56323-6fa9-3386-8a01-6480fb634889>