Fix the issue which seeking is not finished in HLS
SocketInputStream.read() is blocked if it is called with same timing
from another thread. This is called when MediaHTTPConnection.readAt()
and MediaHTTPConnection.disconnect() is called. And disconnect() is
called on LiveSession Looper but readAt() is called on Featcher looper
in PlaylistFetcher. So this fix changes to call also disconnect() on
Fetcher looper not to run read() with same timing.

Bug: 119900000
Test: check HLS behavior manually from stability perspective

Change-Id: Ib8c022af0649fc46ace770d26d0d37b53504139f