Update patch set 2
Patch Set 2: Code-Review-1

It would be better to have an assert when an unknown source is present than a warning.
The configuration file should  be correct and there is no reason to be resilient to programming errors in this file.

Also note that the .conf format is obsolete and you should move to the xml format for new designs.

Patch-set: 2
Reviewer: Gerrit User 1003259 <1003259@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1