Update patch set 2
Patch Set 2: -Code-Review

Hi Eric!

> It would be better to have an assert when an unknown source is present than a warning.

The warning can be easily turned into error here, but the problem is that the crashed function is IN FACT checking if a keyword is in an array of acceptable values. There are no prior checks of file inputs as shown in stacktrace:

00 AudioPolicyEffects::inputSourceNameToEnum(
        const char *name)

01 AudioPolicyEffects::loadInputEffectConfigurations(
        cnode *root, Vector <EffectDesc *>& effects)

02 AudioPolicyEffects::loadAudioEffectConfig(
        const char *path)

03 AudioPolicyEffects::AudioPolicyEffects(
        AudioPolicyService *audioPolicyService)

> The configuration file should be correct and there is no reason to be resilient to programming errors in this file.

The requirement that the file must be correct does not abandon the necessity of fixing the process segmentation fault.

> Also note that the .conf format is obsolete and you should move to the xml format for new designs.

The XML implementation has slightly different declaration of that array in question:

https://github.com/LineageOS/android_frameworks_av/blob/acdef2d508aad336e6581947bc8bdf1fbf9540fc/media/libeffects/config/src/EffectsConfig.cpp#L108

and I asked a colleague at LineageOS team to verify if there is no crash using audio_effects.xml. Still, AOSP Nougat has no support for audio effects XML configuration as it was added in Android 8.1.0_r1.

Patch-set: 2
Reviewer: Gerrit User 1323125 <1323125@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0