media.metrics: improve thread-safety
There was a situation when many clients call MediaAnalyticsService::submit,
and and the time interval more than kNewSetIntervalNs. The function of
newSummarizerSet may be called for many times, who will change mSummarizerSets
and mCurrentSet.
With one of our device, dead loop happened once, three thread block in the while
loop, which consume a lot CPU resource.
At the same time, mSummarizerSets and mCurrentSet should be protected with mLock,
to improve thread-safety.

Test: NA

Change-Id: Ib91ceedf22d879715da06cd0d69b07d0bc87590a
Signed-off-by: zengjing <zengjing@xiaomi.com>