fix mediaserver crash when extractVideoFrame
In some situation, the first frame may be obtain without INFO_FORMAT_CHANGED.
Then outputFormat was null, and mediaserver abort.
In our device, we found some times hardware decoder decode H264 video fail,
the function getFrameAtTime will try to decode it with google avc decoder.
Then the situation occured which result in crash of mediaserver.
It is ok with Android N, for outputFormat was check before use.
To fix it, try to check and get outputFormat before use it.

Test: NA

Change-Id: I8f058713e01d029e465482bb6ef682318e9a273f
Signed-off-by: zengjing <zengjing@xiaomi.com>