Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > If it is OK for you I can merge this directly to our internal master and fix the conflict there.
> > let me know?
> 
> Hi Eric,
> 
> I am sorry, because we can not access https://android-build.googleplex.com/builds, so I can not see the conflict.It's glad if you can show us the conflict. Thanks a lot!

I already merged this internally, so you do not need to resolve the conflict.
If you really need this in AOSP we can wait for the internal merge to propagate and see if this conflict then resolves as an empty merge.

Patch-set: 2
Reviewer: Gerrit User 1003259 <1003259@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0