add getUnderrunCount for StaticAudioTrackClientProxy
The underrun frame counter only exists in AudioTrackSharedStreaming.
When getUnderrunCount() of StaticAudioTrackClientProxy was called,
mCblk->u.mStreaming.mUnderrunCount will be return, which could be an
invalid value as 0xFFFFFFFF. In AudioTrack::restoreTrack_l, the function
of getUnderrunCount_l will be called and the return value will be set for
mUnderrunCountOffset. If getUnderrunCount_l was called twice for track with
sharedBuffer, mUnderrunCountOffset will be the sum of 0xFFFFFFFF and
0xFFFFFFFF. Then, overflow happened which result in native crash of APP.
Support getUnderrunCount for StaticAudioTrackClientProxy will resolve it.

Test: NA

Change-Id: Ied9279b398b8bda6bfe26b6b2937deab7d2cd4ba
Signed-off-by: zengjing <zengjing@xiaomi.com>