Update patch set 2
Patch Set 2: Verified+1

We could reproduce this issue in our device with the app of Hill Climb Racing(v1.37.0). 
1、Running the game in PhoneA with bluetooth earphone connected;
2、Call PhoneA with another PhoneB, and answer it.
3、Hang up from PhoneA, and click the button of game to resume
4、Then native crash occur
After merge this change, this issue will not happened.
This issue could not be reproduce with Pixel 2(Android P preview), for restoreTrack_l will only be called for once. May be it is the different of telephone APP result in this differences.
But this issue is obvious, overflow will occur with restoreTrack_l be called for twice.

Patch-set: 2
Reviewer: Gerrit User 1143570 <1143570@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1