Update patch set 3
Patch Set 3:

Does this need to be merged now?
This will conflict with some refactoring in audio policy manager in master and will not apply cleanly.
This particular problem is addressed by the refactoring.
If this does not need to be fixed before next release then we should not merge in AOSP.

Patch-set: 3
Reviewer: Gerrit User 1003259 <1003259@85c56323-6fa9-3386-8a01-6480fb634889>