Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I don't understand this change. If the constructor hasn't finished, how can you even have a pointer to the object in question?

In the constructor of AudioEffect, the set method is called. And in the set method, the
pointer of 'this' is passed to EffectClient to new an instance, in Line 130 of AudioEffect.cpp:
->  mIEffectClient = new EffectClient(this);

And the mIEffectClient is passed to AudioFlinger to create an effect, in Line 134 of AudioEffect.cpp:
-> iEffect = audioFlinger->createEffect((effect_descriptor_t *)&mDescriptor,
-> 			mEffectClient, prority, io, mSessionId, mOpPackageName, mClientPid,
->			&mStatus, &mId, &enabled);

In AudioFlinger::createEffect, the binder proxy of mEffectClient is passed to Threads::createEffect_l,
in Line 3140 of AudioFlinger.cpp:
-> handle = thread->createEffect_l(client, effectClient, priority, sessionId, 
->			&desc, enabled, &lStatus, pinned);

In Threads::createEffect_l, the binder proxy of mEffectClient is used to new an instance of EffectHandle,
in Line 1341 of Threads.cpp:
-> handle = new EffectHandle(effect, client, effectClient, priority);

The EffectHandle's three methods(::setControl, ::commandExecuted, ::setEnabled) will invoke the methods of
EffectClient by binder. For example ::setControl in Line 1772 of Effects.cpp:
-> void AudioFlinger::EffectHandle::setControl(bool hasControl, bool signal, bool enabled)
-> {
-> 		...
->		if (signal && mEffectClient != null) {
->			mEffectClient->controlStatusChanged(hasControl);
->		}
->		...
-> }

See the definition of EffectClient::controlStatusChanged, in line 440 of AudioEffect.h:
-> virtual void controlStatusChanged(bool controlGranted) {
-> 		sp<AudioEffect> effect = mEffect.prompt();
-> 		...
-> }
In the method of controlStatusChanged, we get a sp instance of AudioEffect by wp's prompt, that means the refCount
of AudioEffect will plus 1, and after the method the refCount of AudioEffect will minus 1. And, if the controlStatusChanged
be invoked during the AudioEffect's constructor, no other sp instance holds the object of AudioEffect, so when the 
method of controlStatusChanged finished, the refCount of AudioEffect will be 0, that leads to deleting of AudioEffect.

Patch-set: 2
CC: Gerrit User 1353159 <1353159@85c56323-6fa9-3386-8a01-6480fb634889>