Update patch set 2
Patch Set 2:

So the underlying problem is that a partially constructed object is passed to other parts of the framework, which could then use it while still in this partially constructed state, potentially even deleting it while the constructor is still running.

I'm not familiar with this code, but could it be refactored so that the constructor just creates the basic object, but doesn't initialize it (i.e. doesn't call set()), and then audioflinger or EffectClient call set() to initialize the AudioEffect?

Patch-set: 2
Reviewer: Gerrit User 1001525 <1001525@85c56323-6fa9-3386-8a01-6480fb634889>