audio: update recordingConfiguration when destroyed
RecordingActivityMonitor stores recordingConfiguration by the event send from audioserver
when AudioSession called onRecordingConfigurationUpdate. The RECORD_CONFIG_EVENT_STOP will
be called when changeActiveCount(-1) is called by AudioPolicyManager::stopInput.
If AudioSession was destroyed without AudioPolicyManager::stopInput called, the server of
RecordingActivityMonitor will not update state of this session, and there always be an recorder
in mRecordConfigs.
When switching between SCO and MIC, AudioPolicyManager::checkInputsForDevice is called, and it
causes closeInput() to be called without stopInput(). Then this issue occure.
When destructor is called, if the session was active, we should update recordingConfiguration.

Test: NA

Change-Id: I9c86fad987de1c57b5fc05b8413a8bb695640c1d
Signed-off-by: zengjing <zengjing@xiaomi.com>