audio: update recordingConfiguration when destructed
RecordingActivityMonitor store recordingConfiguration by the event send from audioserver
when AudioSession called onRecordingConfigurationUpdate. The RECORD_CONFIG_EVENT_STOP will
be called when changeActiveCount(-1) was called by AudioPolicyManager::stopInput.
If AudioSession was destructed without AudioPolicyManager::stopInput called, the server of
RecordingActivityMonitor will not update state of this session, and there always be an recorder
in mRecordConfigs without remove.
When switch between SCO and MIC, AudioPolicyManager::checkInputsForDevice will be called when
closeInput() was called without stopInput(). Then this issue occure.
We could reproduct this issue with some game like "Playerunknown's Battlegrounds". Open the MIC
with teammate, connect with Bluetooth. When close the Bluetooth, the input will switch from SCO
to MIC. In this situation, there will be an audiorecord in recording state which was store in
RecordingActivityMonitor.
When destructor was called, if the session was active, we should update recordingConfiguration

Test: NA

Signed-off-by: zengjing <zengjing@xiaomi.com>