audio: update recordingConfiguration when destroyed
RecordingActivityMonitor stores recordingConfiguration by the event send from audioserver
when AudioSession called onRecordingConfigurationUpdate. The RECORD_CONFIG_EVENT_STOP will
be called when changeActiveCount(-1) is called by AudioPolicyManager::stopInput.
If AudioSession was destroyed without AudioPolicyManager::stopInput called, the server of
RecordingActivityMonitor will not update state of this session, and there always be an recorder
in mRecordConfigs.
When switching between SCO and MIC, AudioPolicyManager::checkInputsForDevice is called, and it
causes closeInput() to be called without stopInput(). Then this issue occure.
When destructor is called, if the session was active, we should update recordingConfiguration.

Test: NA

Change-Id: Ic9ebe4d1f0bc3fdaa9a6940abaf79c481f6a9880
Signed-off-by: zengjing <zengjing@xiaomi.com>