Visualizer: fix native crash when visualizer release
When the destructor of Visualizer is called, setEnabled and setCaptureCallBack
will be called. If we don't grant control (mStatus is ALREADY_EXISTS), setEnabled
will return -17, when mEnabled will not be changed (maybe is true). In this situation,
setCaptureCallBack will return INVALID_OPERATION without requestExit of mCaptureThread.
Then the destructor return, and mCaptureLock will be free when mCaptureThread still running.
As a result, native crash happen in periodicCapture when try to lock mCaptureLock.
Visualizer should not be destroyed before mCaptureThread exit.
Store a strong reference of Visualizer in mCaptureThread, when released, we requestExit of
mCaptureThread. After exit, the reference will be clear, and the destructor will be called.
The backtrace when crashed was shown as follow.

Test: NA

Change-Id: I8d936f0f79540345d3e3675f6129bb942a08e423
Signed-off-by: zengjing <zengjing@xiaomi.com>