Update patch set 1
Patch Set 1:

(1 comment)

Note that internally i've reworked this code so that if there are more than about 50 items to be recovered, we start doing the recovery in a separate thread that comes up for air every 50 further records and eventually terminates once it has recovered enough.

this frees the initial submit() thread sooner so the client is released.
It also means that we won't hold the mutex for too long -- so we don't starve any other
submit() threads that are coming in.

Patch-set: 1
Reviewer: Gerrit User 1108919 <1108919@85c56323-6fa9-3386-8a01-6480fb634889>