mediaextractor: Enable seccomp filter for mediaextractor on x86_64
This is used to address a cts testcase failure. This testcase
checks if a seccomp filter is installed for mediaextractor or
not, if not, then fail.

This testcase failed on x86_64 platforms. These is indeed a
seccomp policy file for mediaextractor on x86_64, but it's not
enabled in Android.mk for x86_64 platforms, we need to enable
it in Android.mk for x86_64. And to make mediaextractor work
normally with the seccomp filter enabling, additional system
calls should be added to the seccomp policy files for x86_64.

[bugzilla id]:
117960171

[Android Version]:
android_p_mr0_r0

[Kernel Version]:
4.19.0-rc8

[CTS Version]:
cts-9.0_r1

[Failed Testcase]:
android.security.cts.ProcessMustUseSeccompTest#testMediaextractorHasSeccompFilter

[Error Log]:
10-19 14:06:01 I/ModuleListener:
[1/1] android.security.cts.ProcessMustUseSeccompTest#testMediaextractorHasSeccompFilter
fail: junit.framework.AssertionFailedError: media.extractor must have a seccomp
filter enabled.
The "Seccomp" field of media.extractor's /proc/2847/status file should be set to "2"

[Test Result With This Patch]:
This testcase can pass and mediaextrctor work normally with this patch.

Signed-off-by: ji, zhenlong z <zhenlong.z.ji@intel.com>