Framework Audio: Mobile phone open log system
The context log can be clearly seen, and another thread is also operating on mOutputRoutes.Here is the remove operation, AudioPolicyManager::getOutputForAttr.The place where getDevicesForStream is called does not have lock protection.

At present, try to add a lock in the AudioPolicyManager, protect mOutputRoutes (mInputRoutes is not protected for the time being), and incorporate the version tracking test.

Test:CtsLiblogTestCasts CtsMediaTestCases GtsMediaTestCases

Test: manual
1.Fuzz test did not appear collapse
2.Monkey test did not appear abnormal
3.Play media files normally, no abnormalities

Change-Id: I9020cc8a64101b26f55d8b5710b68e48670852b7