Update patch set 1
Patch Set 1:

I'm not totally clear on how this script works - if we have a mismatch in the map and what's in the library, does the build then break with some sort of printout of the mismatch?

In any case, sounds good to me - I assume no mismatch was found to start with?

Patch-set: 1
Reviewer: Gerrit User 1010050 <1010050@85c56323-6fa9-3386-8a01-6480fb634889>