Update patch set 1
Patch Set 1:

in the x86_64 policy, it's grouped with several other syscalls under a "required by addresssantizer" comment -- it'd be good to structure this addition similarly.

otherwise -- I'm ok with adding it; it makes the arm64 policy consistent the other 3 policies
in this directory.

Patch-set: 1
Reviewer: Gerrit User 1108919 <1108919@85c56323-6fa9-3386-8a01-6480fb634889>