Add mediaextractor.policy in seccomp filter for x86_64
This is used to address a cts testcase failure. This testcase
checks if a seccomp filter is installed for mediaextractor or
not, if not, then fail.

This testcase failed on x86_64 platforms. These is indeed a
seccomp policy file for mediaextractor on x86_64, but it's not
enabled in Android.mk for x86_64 platforms, we need to enable
it in Android.mk for x86_64. And to make mediaextractor work
normally with the seccomp filter enabling, additional system
calls should be added to the seccomp policy files for x86_64.

Bug: 117960171
bug: 120238328
Test: run CTS case - android.security.cts.ProcessMustUseSeccompTest

Change-Id: I08ac4d75d77061a48641973cff61d51fee3c5c28
Signed-off-by: ji, zhenlong z <zhenlong.z.ji@intel.com>
Signed-off-by: Yong Yao <yong.yao@intel.com>