Respect config.disable_cameraservice in initrc.
'config.disable_cameraservice' can be used to tell the Camera API not to
connect to cameraserver. When this is set, we'd like cameraserver not to
be started.

Not starting cameraserver saves ~2.9 MiB of RAM, which is valuable on
some low-ram Android TV devices.

This depends on I9cce3e4dc2b57339cdc44830e19aa0218b8282c3 to provide a
default value, which I will submit first.

Bug: 117139498
Bug: 118369095
Test: Flash device, cameraserver comes up. Set
	config.disable_cameraserver=1 (through /data/local.prop), reboot,
	cameraserver doesn't come up.
Change-Id: I20cc7a956c561517b340d8b69382de713dda74e3