Adjust the extractor plugin loading sequence
Skia does the HeifImage decoding and creates SkStream
which is then passed to the extractor via setDataSource.
But the SkStream is a wrapper of Java Input Stream which
does not support rewind.

If other extractor sniffing has triggered the Java InputStream
throws EOF exception, when MPEG4Extractor begins to read at
offset 0, it would fail.

Fix: make sure MP4 Extractor is the first item in the List.

Bug: 115948759

Test: push a heif image of small size and reboot device. Try to
view it in Gallery.

Change-Id: Ic379c2c69cdef42c25f4e9b95082b5b740daf229
Signed-off-by: Dan Liang <dan.liang@intel.com>
Signed-off-by: Zhiquan Liu<zhiquan.liu@intel.com>