Update patch set 1
Patch Set 1:

I don't think this CL is the right fix, making sure mp4 is the first doesn't solve the problem, it'll fail again when the file is of other formats.


We already had fixes merged for similar crashes in MediaMetadataRetriever._setDataSource. The fix is in internal master, and is also cherry-picked to AOSP (see b/111897855, b/117625929):

https://android-review.googlesource.com/c/platform/frameworks/base/+/796718

Patch-set: 1
Reviewer: Gerrit User 1039918 <1039918@85c56323-6fa9-3386-8a01-6480fb634889>