Update patch set 5
Patch Set 5:

Makes sense.

I'd try with 

on property:config.disable_cameraservice=1
    stop cameraserver

to see if that does what you need.  That at least makes it so the default value doesn't need to be provided for all non-ATV devices.

Otherwise, this is fine and all as is.  Just a little bit non-ideal.

Patch-set: 5