Update patch set 1
Patch Set 1: Code-Review+2

(2 comments)

> Patch Set 1:
> 
> mast@: A copy and paste of similar changes you made. Any other changes needed elsewhere to make it work?

Well, I noticed some flapping in drmserver and cameraserver while playing around with moving libs to the Runtime APEX - what I saw there is that they fail to start because a library is missing, and so init will retry them. In a normal boot that's not a showstopper, but tests might be more sensitive to service startup errors (and rightly so).

I did http://r.android.com/880332 in my WIP topic for this, but since I reckon it's random which services fail due to the race I just added "updatable" to everything in frameworks/av. Feel free to patch it in here. I doubt it's enough though; we should invert the flag so that we don't need to add it in all these places - filed b/123275379 to track that.

Patch-set: 1
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2