Mark AudioPolicyInterface::setRecordSilenced() as pure virtual
Bug: http://b/116873221

If not, Clang assumes that the TU which defines this function will
contain the VTable for this class.  Since there's no out-of-line
definition of this funciton, no TU ends up with the VTable.

This causes a problem with coverage builds, which are built with -O0,
where calls don't get inlined, thereby requiring a definition of the
VTable.  For regular builds, the VTable is not required since the
virtual calls get inlined/optimized-out.

Test: Build with and without -O0.
Change-Id: Ibb54cae9b421a975d71734deff57db133e8eb993