Update patch set 1
Patch Set 1:

What does it mean to be double loadable ? From Android.bp:

whether this module is capable of being loaded with other instance (possibly an older version) of the same module in the same process. Currently, a shared library that is a member of VNDK (vndk: {enabled: true}) can be double loaded in a vendor process if the library is also a (direct and indirect) dependency of an LLNDK library. Such libraries must be explicitly marked as `double_loadable: true` by the owner, or the dependency from the LLNDK lib should be cut if the lib is not designed to be double loaded.

But I do not understand. Is it that the library is loaded twice (how do the symbol not conflict ?) or does it mean the library is only loaded once and the global shared ?

Patch-set: 1
Reviewer: Gerrit User 1136369 <1136369@85c56323-6fa9-3386-8a01-6480fb634889>