Use libnativewindow's AHardwareBuffer directly
This is a part of cutting dependency from libmediandk to
libandroid_runtime.

The following function is exported from libandroid_runtime.
* android_hardware_HardwareBuffer_convertToGrallocUsageBits

This function just delegates to libnativewindow's
AHardwareBuffer_convertToGrallocUsageBits function.

This commit is for libmediandk to call libnativewindow's
function directly in order to cut the dependency to libandroid_runtime.

Bug: 124268753
Test: m -j
Change-Id: I5f927fc26829031a8c91a7ba51178a34d737a953
Merged-In: I5f927fc26829031a8c91a7ba51178a34d737a953