Update patch set 2
Patch Set 2:

interesting that the countdown itself is considered as a shutter sound by the regulator.
Are we sure of that or is it an interpretation of the tester?
do we still hear the shutter sound at the end of the countdown?

Patch-set: 2
Reviewer: Gerrit User 1003259 <1003259@85c56323-6fa9-3386-8a01-6480fb634889>