Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> interesting that the countdown itself is considered as a shutter sound by the regulator.
> Are we sure of that or is it an interpretation of the tester?
> do we still hear the shutter sound at the end of the countdown?

I do not know if the count down for self timer is actually included in the regulation but I guess it's up to the interpretation of each vendor. Anyway our products are using system enforced playback for count down as well.
When the issue happens, both count down and shutter sound can not be heard.

Do you think there is an issue in this CL?

Patch-set: 2
Reviewer: Gerrit User 1243029 <1243029@85c56323-6fa9-3386-8a01-6480fb634889>