Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> majiajun@, it seems like you are moving the loading of sound from camera start to playSound.
> Would it cause latency increase while taking a picture, or start video recording when shutter sound are played?


We evaluated the negative impact of the modification, which affects the time when the audio file is played for the first time. It takes 20ms in a single time (only one file is loaded at a time), and has no effect on the user's use of taking pictures and recording video scenes. But it saves memory and speeds up camera startup for more than 60ms, especially on low-end machines. This modification has been on Huawei phones for many years. After many rounds of commercial iterations, the code may not be standardized, but the idea should be feasible.

Patch-set: 2
Reviewer: Gerrit User 1418717 <1418717@85c56323-6fa9-3386-8a01-6480fb634889>