Update patch set 2
Patch Set 2:

> We evaluated the negative impact of the modification, which affects the time when the audio file is played for the first time. It takes 20ms in a single time (only one file is loaded at a time), and has no effect on the user's use of taking pictures and recording video scenes. But it saves memory and speeds up camera startup for more than 60ms, especially on low-end machines. This modification has been on Huawei phones for many years. After many rounds of commercial iterations, the code may not be standardized, but the idea should be feasible.

majiajun@, does this have no affect on shutter usage because the sounds are loaded in parallel to the camera operating? Otherwise, wouldn't this increase camera delay when the shutter is used for the first time?

Shuzhen, if you are reluctant to merge this, what are you thoughts on putting this change behind the lazy HAL flag?

Patch-set: 2
Reviewer: Gerrit User 1321491 <1321491@85c56323-6fa9-3386-8a01-6480fb634889>