Update patch set 2
Patch Set 2: Code-Review-1

This has already been converted to .bp in internal master.  

It's fine if you want to cherry-pick that back to AOSP right now instead of waiting for it, but I don't think we should have two different .bp files around.

Plus this is putting a copyright header in the middle of a file, which is weird.

Patch-set: 2
Reviewer: Gerrit User 1010050 <1010050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1