Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Now recording sound plays unconditionally whenever camera recording starts/stops through Camera Service. Since VT solution uses Stagefright Recorder, we could not control enabling/disabling these sounds without adding an extra API. This patch has been made so that user does not hear sound every time a VT call is started/stopped. Is there any other way to achieve the same functionality without adding extra API?

The preferred way is to use the non-deprecated Camera2 API for VT, which is what we've told other VT providers.  

I realize that's a large rewrite of the existing solution, but the old camera API has been deprecated for over 4 years and we are generally not adding features to it.

Patch-set: 3
Reviewer: Gerrit User 1010050 <1010050@85c56323-6fa9-3386-8a01-6480fb634889>